Data Association
Refactored the association of data between users, comments and campsites.

Each Comment and Campsite now has a reference to the ID and username of the user that created it.
Refactored Campsite routes to include protection middleware to prevent anonymous users from creating new Campsites or being able to access the "New Campsite" submission form.