Campsite Authorisation
Implemented middleware to check and allow for user authorisation of campsites. This means that only the Author/Creator of a Campsite is able to modify or delete the campsite.
Middleware checking for all related RESTful routes - edit, update & destroy.

*** Additional Changes ***
Refactored Google Maps from EJS partial to simply be embedded into the Campsite show page.
Currently hardcoded lat and lng data but is setup to allow pulling data from the Campsite once the Campsite model is update.